Updated Int to long in InPlayerSubscription

+++ Data/src/main/java/com/sdk/data/model/subscription/SubscriptionModel.kt
@@ -8,13 +8,13 @@ data class SubscriptionModel(
         val asset_id: Int,
         val asset_title: String,
         val cancel_token: String,
-        val created_at: Int,
+        val created_at: Long,
         val currency: String,
         val description: String,
         val formatted_amount: String,
         val merchant_id: Int,
-        val next_billing_date: Int,
+        val next_billing_date: Long,
         val status: String,
         val unsubscribe_url: String,
-        val updated_at: Int
+        val updated_at: Long
 )
